backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-pdf-view==3.1.3
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pdfkit==1.0.0
psycopg==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
